[jax2tf] separated the primitives_test.py
