+++ jax/experimental/jax2tf/tests/primitives_test.py
@@ -11,7 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Tests for the jax2tf transformation."""
+"""Tests for JAX primitive coverage."""
 
 import unittest
 
@@ -106,7 +106,7 @@ INDEX = (
 )
 
 
-class TfOpsTest(tf_test_util.JaxToTfTestCase):
+class JaxPrimitiveTest(tf_test_util.JaxToTfTestCase):
 
   def test_primitive_coverage(self):
     """Fail if there are JAX primitives that are not implemented."""
@@ -127,47 +127,6 @@ class TfOpsTest(tf_test_util.JaxToTfTestCase):
         self.assertNotIn(p, tf_impl)  # Should not be in both tf_impl and tf_not_yet_impl
       else:
         self.assertIn(p, tf_impl)
-
-  def test_basics(self):
-    f_jax = lambda x: jnp.sin(jnp.cos(x))
-    _, res_tf = self.ConvertAndCompare(f_jax, 0.7)
-    self.assertIsInstance(res_tf, tf.Tensor)
-
-  def test_variable_input(self):
-    f_jax = lambda x: jnp.sin(jnp.cos(x))
-    f_tf = jax2tf.convert(f_jax)
-    v = tf.Variable(0.7)
-    self.assertIsInstance(f_tf(v), tf.Tensor)
-    self.assertAllClose(f_jax(0.7), f_tf(v))
-
-  def test_jit(self):
-    f_jax = jax.jit(lambda x: jnp.sin(jnp.cos(x)))
-    self.ConvertAndCompare(f_jax, 0.7)
-
-  def test_nested_jit(self):
-    f_jax = jax.jit(lambda x: jnp.sin(jax.jit(jnp.cos)(x)))
-    f_tf = jax2tf.convert(f_jax)
-    np.testing.assert_allclose(f_jax(0.7), f_tf(0.7))
-
-  def test_converts_jax_arrays(self):
-    f_tf = tf.function(lambda x: x)
-    self.assertEqual(f_tf(jnp.zeros([])).numpy(), 0.)
-    self.assertEqual(f_tf(jnp.ones([])).numpy(), 1.)
-    f_tf = tf.function(lambda x: x + x)
-    self.assertEqual(f_tf(jnp.ones([])).numpy(), 2.)
-
-    # Test with ShardedDeviceArray.
-    n = jax.local_device_count()
-    mk_sharded = lambda f: jax.pmap(lambda x: x)(f([n]))
-    f_tf = tf.function(lambda x: x)
-    self.assertAllClose(f_tf(mk_sharded(jnp.zeros)).numpy(),
-                        np.zeros([n]))
-    self.assertAllClose(f_tf(mk_sharded(jnp.ones)).numpy(),
-                        np.ones([n]))
-
-  def test_function(self):
-    f_jax = jax.jit(lambda x: jnp.sin(jnp.cos(x)))
-    self.ConvertAndCompare(f_jax, 0.7, with_function=True)
 
   @parameterized.named_parameters(jtu.cases_from_list(
     dict(testcase_name=f"_{f_jax.__name__}",
@@ -382,15 +341,6 @@ class TfOpsTest(tf_test_util.JaxToTfTestCase):
                     ]:
       self.ConvertAndCompare(f_jax, rng_key, with_function=True)
 
-  def test_gradients_disabled(self):
-    f = jax2tf.convert(jnp.tan)
-    x = tf.ones([])
-    with tf.GradientTape() as tape:
-      tape.watch(x)
-      y = f(x)
-    with self.assertRaisesRegex(ValueError,
-                                'jax2tf currently does not support gradients'):
-      tape.gradient(y, x)
 
   def test_zeros_like(self):
     v = np.float32(2.)
@@ -400,19 +350,6 @@ class TfOpsTest(tf_test_util.JaxToTfTestCase):
   def test_stop_gradient(self):
     f = jax2tf.convert(lax.stop_gradient)
     self.assertEqual(f(tf.ones([])), 1.)
-
-  def test_checkpoint_wrapper_types(self):
-    m = tf.Module()
-    m.a = [tf.Module(), tf.Module()]
-    m.b = (tf.Module(), tf.Module())
-    m.c = {'a': tf.Module(), 'b': tf.Module()}
-    self.assertNotEqual(type(m.a), list)
-    self.assertNotEqual(type(m.b), tuple)
-    self.assertNotEqual(type(m.c), dict)
-    self.assertLen(jax.tree_leaves(m.a), 2)
-    self.assertLen(jax.tree_leaves(m.b), 2)
-    self.assertLen(jax.tree_leaves(m.c), 2)
-
 
 if __name__ == "__main__":
   absltest.main()
